create common module
